+++ faked-arduino/seasim/src/seasim.c
@@ -42,7 +42,7 @@ seasim_get_generic_output(uint8_t pin, uint8_t pin_type)
 }
 
 uint8_t 
-seasim_set_generic_input(uint8_t pin, uint8_t val, uint8_t pin_type)
+seasim_set_generic_input(uint8_t pin, unsigned int val, uint8_t pin_type)
 {
   return ext_set_generic_input(pin, val, pin_type);
 }
